Fixed JS syntax errors on cookiePath
Currrently, if the cookiePath is null as per the configuration, we get a SyntaxError ("Unexpected token }" or "expected expression, got '}'" depending on the browser) and if it is defined (to '/' for example as in the doc) we have another SyntaxError ("invalid escape sequence" or "Invalid or unexpected token" depending on the browser). Simply escaping the value solves both cases for me.